refactor: normalize OwncastLogo component
