+++ web/components/common/index.ts
@@ -1,2 +1,1 @@
 export { default as UserDropdown } from './UserDropdown';
-export { default as OwncastLogo } from './Logo';
